+++ src/bios/display_text.h
@@ -9,9 +9,9 @@
 void InitDisplayText(BIOSState* bios);
 
 // Read a byte from the display text buffer.
-uint8_t ReadDisplayTextByte(BIOSState* bios, uint16_t address);
+uint8_t ReadDisplayTextByte(BIOSState* bios, uint32_t address);
 // Write a byte to the display text buffer.
-void WriteDisplayTextByte(BIOSState* bios, uint16_t address, uint8_t value);
+void WriteDisplayTextByte(BIOSState* bios, uint32_t address, uint8_t value);
 
 #endif  // YAX86_IMPLEMENTATION
